+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/api/ProcessesResources.scala
@@ -30,6 +30,7 @@ import pl.touk.nussknacker.ui.process.deployment.DeploymentService
 import pl.touk.nussknacker.ui.process.marshall.ProcessConverter
 import pl.touk.nussknacker.ui.process.repository.FetchingProcessRepository
 import pl.touk.nussknacker.ui.process.repository.FetchingProcessRepository.FetchProcessesDetailsQuery
+import pl.touk.nussknacker.ui.process.repository.ProcessRepository.RemoteUserName
 import pl.touk.nussknacker.ui.security.api.LoggedUser
 import pl.touk.nussknacker.ui.uiresolving.UIProcessResolving
 import pl.touk.nussknacker.ui.util._
@@ -146,14 +147,16 @@ class ProcessesResources(
               }
             } ~ (put & canWrite(processId)) {
               entity(as[UpdateProcessCommand]) { updateCommand =>
-                complete {
-                  processService
-                    .updateProcess(processId, updateCommand)
-                    .withSideEffect(response => sideEffectAction(response.toOption.flatMap(_.processResponse)) { resp =>
-                      OnSaved(resp.id, resp.versionId)
-                    })
-                    .map(_.map(_.validationResult))
-                    .map(toResponseEither[ValidationResult])
+                canOverrideUsername(processId.id, updateCommand.forwardedUserName)(ec, user) {
+                  complete {
+                    processService
+                      .updateProcess(processId, updateCommand)
+                      .withSideEffect(response => sideEffectAction(response.toOption.flatMap(_.processResponse)) { resp =>
+                        OnSaved(resp.id, resp.versionId)
+                      })
+                      .map(_.map(_.validationResult))
+                      .map(toResponseEither[ValidationResult])
+                  }
                 }
               }
             } ~ (get & skipValidateAndResolveParameter) { skipValidateAndResolve =>
@@ -193,15 +196,19 @@ class ProcessesResources(
           }
         } ~ path("processes" / Segment / Segment) { (processName, category) =>
           authorize(user.can(category, Permission.Write)) {
-            parameter(Symbol("isSubprocess") ? false) { isSubprocess =>
-              post {
-                complete {
-                  processService
-                    .createProcess(CreateProcessCommand(ProcessName(processName), category, isSubprocess))
-                    .withSideEffect(response => sideEffectAction(response) { process =>
-                      OnSaved(process.id, process.versionId)
-                    })
-                    .map(toResponseEither[ProcessResponse](_, StatusCodes.Created))
+            optionalHeaderValue(RemoteUserName.extractFromHeader) { remoteUserName =>
+              canOverrideUsername(category, remoteUserName)(user) {
+                parameter(Symbol("isSubprocess") ? false) { isSubprocess =>
+                  post {
+                    complete {
+                      processService
+                        .createProcess(CreateProcessCommand(ProcessName(processName), category, isSubprocess, remoteUserName))
+                        .withSideEffect(response => sideEffectAction(response) { process =>
+                          OnSaved(process.id, process.versionId)
+                        })
+                        .map(toResponseEither[ProcessResponse](_, StatusCodes.Created))
+                    }
+                  }
                 }
               }
             }